+++ lib/lightstreamer/subscription.rb
@@ -19,16 +19,17 @@ module Lightstreamer
     # @return [Array]
     attr_reader :fields
 
-    # The operation mode of this subscription.
+    # The operation mode of this subscription. The four supported operation modes are: `:command`, `:distinct`, `:merge`
+    # and `:raw`. See the Lightstreamer documentation for details on the different modes.
     #
-    # @return [:distinct, :merge]
+    # @return [:command, :distinct, :merge, :raw]
     attr_reader :mode
 
     # The name of the data adapter from the Lightstreamer session's adapter set that should be used, or `nil` to use the
     # default data adapter.
     #
     # @return [String, nil]
-    attr_reader :adapter
+    attr_reader :data_adapter
 
     # The selector for table items.
     #
@@ -37,7 +38,8 @@ module Lightstreamer
 
     # The maximum number of updates this subscription should receive per second. If this is set to zero, which is the
     # default, then there is no limit on the update frequency. If set to `:unfiltered` then unfiltered streaming will be
-    # used for this subscription and it is possible for overflows to occur (see {#on_overflow}).
+    # used for this subscription and it is possible for overflows to occur (see {#on_overflow}). If {#mode} is `:raw`
+    # then the maximum update frequency is treated as `:unfiltered` regardless of its actual value.
     #
     # @return [Float, :unfiltered]
     attr_reader :maximum_update_frequency
@@ -59,7 +61,7 @@ module Lightstreamer
       @items = options.fetch(:items)
       @fields = options.fetch(:fields)
       @mode = options.fetch(:mode).to_sym
-      @adapter = options[:adapter]
+      @data_adapter = options[:data_adapter]
       @selector = options[:selector]
       @maximum_update_frequency = sanitize_frequency options[:maximum_update_frequency]
 
@@ -91,13 +93,15 @@ module Lightstreamer
     #                 subscription's {#mode} is `:distinct` then `:snapshot` can also be an integer specifying the
     #                 number of events the server should send as part of the snapshot. If this latter option is used
     #                 then any callbacks registered with {#on_end_of_snapshot} will be called once the snapshot for each
-    #                 item is complete.
+    #                 item is complete. This option is ignored when {#mode} is `:raw`.
     def start(options = {})
-      session.control_request(*start_control_request_args(options)) unless @active
+      return if @active
+
+      session.control_request(*start_control_request_args(options))
       @active = true
     end
 
-    # Returns the arguments to pass to to {Session#control_request} in order ot start this subscription with the given
+    # Returns the arguments to pass to to {Session#control_request} in order to start this subscription with the given
     # options.
     #
     # @param [Hash] options The options to start the subscription with.
@@ -106,8 +110,8 @@ module Lightstreamer
     def start_control_request_args(options = {})
       operation = options[:silent] ? :add_silent : :add
 
-      options = { LS_table: id, LS_mode: mode.to_s.upcase, LS_id: items, LS_schema: fields, LS_data_adapter: adapter,
-                  LS_requested_max_frequency: maximum_update_frequency, LS_selector: selector,
+      options = { LS_table: id, LS_mode: mode.to_s.upcase, LS_id: items, LS_schema: fields, LS_selector: selector,
+                  LS_data_adapter: data_adapter, LS_requested_max_frequency: maximum_update_frequency,
                   LS_snapshot: options.fetch(:snapshot, false) }
 
       [operation, options]
@@ -129,7 +133,8 @@ module Lightstreamer
 
     # Sets this subscription's maximum update frequency. This can be done while a subscription is streaming data in
     # order to change its update frequency limit, but an actively streaming subscription cannot switch between filtered
-    # and unfiltered dispatching, and {TableModificationNotAllowedError} will be raised if this is attempted.
+    # and unfiltered dispatching, and {TableModificationNotAllowedError} will be raised if this is attempted. If {#mode}
+    # is `:raw` then the maximum update frequency is treated as `:unfiltered` regardless of its actual value.
     #
     # @param [Float, :unfiltered] new_frequency The new maximum update frequency. See {#maximum_update_frequency} for
     #        details.
@@ -142,13 +147,42 @@ module Lightstreamer
     # Clears all current data stored for this subscription. New data will continue to be processed as it becomes
     # available.
     def clear_data
-      @data = (0...items.size).map { {} }
+      @data_mutex.synchronize { @data = (0...items.size).map { SubscriptionItemData.new } }
+    end
+
+    # Returns a copy of the current data of one of this subscription's items. If {#mode} is `:merge` then the returned
+    # object will be a hash of the item's state, if it is `:command` then an array of row data for the item will be
+    # returned, and if it is `:distinct` or :raw` then just the most recent update received for the item will be
+    # returned. The return value will be `nil` if no data for the item has been set or received.
+    #
+    # @param [String] item_name The name of the item to return the current data for.
+    #
+    # @return [Hash, Array, nil] A copy of the item data.
+    def item_data(item_name)
+      index = @items.index item_name
+      raise ArgumentError, 'Unknown item' unless index
+
+      @data_mutex.synchronize { @data[index].data && @data[index].data.dup }
+    end
+
+    # Sets the current data for the item with the specified name. This is only allowed when {mode} is `:command` or
+    # `:merge`. Raises an exception if the specified item name or item data is invalid.
+    #
+    # @param [String] item_name The name of the item to set the data for.
+    # @param [Hash, Array<Hash>] item_data The new data for the item. If {#mode} is `:merge` this must be a hash. If
+    #        {#mode} is `:command` then this must be an `Array<Hash>` and each hash entry must have a unique `:key`
+    #        value.
+    def set_item_data(item_name, item_data)
+      index = @items.index item_name
+      raise ArgumentError, 'Unknown item' unless index
+
+      @data_mutex.synchronize { @data[index].set_data item_data, mode }
     end
 
     # Adds the passed block to the list of callbacks that will be run when new data for this subscription arrives. The
     # block will be called on a worker thread and so the code that is run by the block must be thread-safe. The
-    # arguments passed to the block are `|subscription, item_name, item_data, new_values|`. If {#mode} is `:distinct`
-    # then the values of `item_data` and `new_values` will be the same.
+    # arguments passed to the block are `|subscription, item_name, item_data, new_data|`. If {#mode} is `:distinct`
+    # or `:raw` then `item_data` and `new_data` will be the same.
     #
     # @param [Proc] callback The callback that is to be run when new data arrives.
     def on_data(&callback)
@@ -156,8 +190,10 @@ module Lightstreamer
     end
 
     # Adds the passed block to the list of callbacks that will be run when the server reports an overflow for this
-    # subscription. The block will be called on a worker thread and so the code that is run by the block must be
-    # thread-safe. The arguments passed to the block are `|subscription, item_name, overflow_size|`.
+    # subscription. This is only relevant when this subscription's {#mode} is `:command` or `:raw`, or if
+    # {#maximum_update_frequency} is `:unfiltered`. The block will be called on a worker thread and so the code that is
+    # run by the block must be thread-safe. The arguments passed to the block are `|subscription, item_name,
+    # overflow_size|`.
     #
     # @param [Proc] callback The callback that is to be run when an overflow is reported for this subscription.
     def on_overflow(&callback)
@@ -165,42 +201,19 @@ module Lightstreamer
     end
 
     # Adds the passed block to the list of callbacks that will be run when the server reports an end-of-snapshot
-    # notification for this subscription. The block will be called on a worker thread and so the code that is run by the
-    # block must be thread-safe. The arguments passed to the block are `|subscription, item_name|`.
+    # notification for this subscription. End-of-snapshot notifications are only sent when {#mode} is `:command` or
+    # `:distinct` and `snapshot: true` was passed to {#start}. The block will be called on a worker thread and so the
+    # code that is run by the block must be thread-safe. The arguments passed to the block are `|subscription,
+    # item_name|`.
     #
     # @param [Proc] callback The callback that is to be run when an overflow is reported for this subscription.
     def on_end_of_snapshot(&callback)
       @data_mutex.synchronize { @callbacks[:on_end_of_snapshot] << callback }
     end
 
-    # Removes all {#on_data} and {#on_overflow} callbacks present on this subscription.
+    # Removes all {#on_data}, {#on_overflow} and {#on_end_of_snapshot} callbacks present on this subscription.
     def clear_callbacks
       @data_mutex.synchronize { @callbacks = { on_data: [], on_overflow: [], on_end_of_snapshot: [] } }
-    end
-
-    # Returns a copy of the current data of one of this subscription's items.
-    #
-    # @param [String] item_name The name of the item to return the current data for.
-    #
-    # @return [Hash] A copy of the item data.
-    def item_data(item_name)
-      index = @items.index item_name
-      raise ArgumentError, 'Unknown item' unless index
-
-      @data_mutex.synchronize { @data[index].dup }
-    end
-
-    # Sets the current data for the item with the specified name.
-    #
-    # @param [String] item_name The name of the item to set the data for.
-    # @param [Hash] item_data The new data for the item.
-    def set_item_data(item_name, item_data)
-      index = @items.index item_name
-      raise ArgumentError, 'Unknown item' unless index
-
-      raise ArgumentError, 'Item data must be a hash' unless item_data.is_a? Hash
-
-      @data_mutex.synchronize { @data[index] = item_data.dup }
     end
 
     # Processes a line of stream data if it is relevant to this subscription. This method is thread-safe and is intended
@@ -208,8 +221,7 @@ module Lightstreamer
     #
     # @param [String] line The line of stream data to process.
     #
-    # @return [Boolean] Whether the passed line of stream data was relevant to this subscription and was successfully
-    #         processed by it.
+    # @return [Boolean] Whether the passed line of stream data was processed by this subscription.
     #
     # @private
     def process_stream_data(line)
@@ -229,16 +241,10 @@ module Lightstreamer
     def process_update_message(message)
       return unless message
 
-      @data_mutex.synchronize { process_new_values message.item_index, message.values }
-    end
-
-    def process_new_values(item_index, new_values)
-      data = @data[item_index]
-
-      data.replace(new_values) if mode == :distinct
-      data.merge!(new_values) if mode == :merge
-
-      run_callbacks :on_data, @items[item_index], data, new_values
+      @data_mutex.synchronize do
+        @data[message.item_index].send "process_new_#{mode}_data", message.data.dup
+        run_callbacks :on_data, @items[message.item_index], @data[message.item_index].data, message.data
+      end
     end
 
     def process_overflow_message(message)