+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ### 0.8 — Unreleased
 
+- Added support for Lightstreamer's `COMMAND` and `RAW` subscription modes
 - Added support for Lightstreamer's polling mode and switching an active session between streaming and polling
+- `Lightstreamer::Subscription#item_data` now return `nil` if the requested item has never had any data set
+- Renamed `Lightstreamer::Subscription#adapter` to `Lightstreamer::Subscription#data_adapter` and the `--adapter`
+  command-line option to `--data-adapter`
+- The `--mode` command-line option is now required because the default value of `merge` has been removed
 
 ### 0.7 — July 31, 2016